rag-api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
langchain>=0.3.14
langchain-openai>=0.3.0
langchain-chroma>=0.2.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.60
chromadb>=0.5.23
pydantic>=2.10.0
python-multipart>=0.0.18
pytest>=8.3.0
httpx>=0.28.0
pypdf>=5.0.0
